Convert external-project example to full SuperProject example
Note that although this example successfully builds, several variables
are not set by Find_Package.

Although FLTK_FLUID_EXECUTABLE is set to fluid, it is not a full path
and I would be surprised if it actually works.

It might be good to extend this example to one that uses a *.fl file
managed by CMake via FLTK_WRAP_UI.